crewai_integration/finance_crews/requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
Markdown==3.8
python-dotenv==1.1.0
schedule==1.2.2
streamlit==1.45.1
weasyprint==65.1
weave==0.51.54